requirements/dev.txt
albumentations==1.3.0
dvc[s3]==2.13.0
efficientnet-pytorch==0.7.1
fiftyone~=0.19.1
hydra-core==1.3.2
imageio~=2.22.1
imutils==0.5.4
joblib==1.1.0
matplotlib==3.5.3
mlflow==2.1.1
numpy==1.23.2
omegaconf==2.3.0
opencv-python==4.5.5.64
openmim==0.3.6
openpyxl==3.0.10
pandas==1.3.5
Pillow==9.2.0
pre-commit==2.21.0
pretrainedmodels==0.7.4
ptflops==0.6.9
py-cpuinfo==9.0.0
pytest==7.2.0
PyYAML~=6.0
requests~=2.28.1
requests-toolbelt==0.10.1
scikit-image==0.19.3
scikit-learn==1.1.3
scipy==1.9.3
seaborn==0.12.2
setuptools~=67.6.0
six~=1.16.0
supervisely==6.68.10
terminaltables~=3.1.10
timm==0.6.12                     # Linknet and PSPNet are not tested with timm > 0.3.2
toml==0.10.2
tomli==2.0.1
torch==1.13.1                    # Linknet and PSPNet are not working with torch > 1.8.1
torchvision==0.14.1
tqdm==4.64.0
urllib3==1.26.15
